Attempt a monkeypatch for PHP Bug #52144
This may not even be required, as I can't tell whether it's meant to be fixed or not in the first place.
Part of the code from the patch laruence@php.net submitted is already in some places, though not all of it.

Consider this a "just in case", because the patch shouldn't break anything further